requirements.txt
# Core LangChain (Modern Stack - Pydantic 2 Compatible)
langchain>=0.3,<0.4
langchain-openai>=0.2
langchain-anthropic>=0.2
langchain-community>=0.2
langchain-huggingface>=0.0.5

# Embeddings (local, free)
sentence-transformers>=3.4.1
transformers>=4.41,<5.0
huggingface-hub>=0.34

# Vector store
faiss-cpu>=1.12.0

# Document processing
pypdf==3.17.4

# Evaluation
ragas  # Latest stable version
datasets>=2.16

# Web UI
streamlit>=1.29.0

# Observability and metrics
prometheus-client==0.19.0

# Environment management
python-dotenv==1.0.0

# Utilities
tenacity==8.2.3
pydantic>=2,<3
